smart-receipt-scanner/requirements.txt
# Deep Learning & ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentencepiece>=0.1.99
scikit-learn>=1.3.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data & Storage
pandas>=2.0.0
numpy>=1.24.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dateutil>=2.8.0
